renderHTML: Expand dictionary to provide color contrast report options
Each of the existing options now has a .skip-aa alternative that skips the AA of the report table, a .skip-aaa alternative that skips the AAA of the report table, and a .skip-table alternative that skips the table altogether. For example, in the design system I'm currently using this for, I only use the pairs that meet WCAG AAA standards for large text; since I'm using the highest standard, the table is redundant. That the pair is listed at all means that it passes all criteria.